Remove use of AnnData constructor dtype kwarg from normalization and preprocessing tests

+++ scanpy/tests/test_preprocessing.py
@@ -152,7 +152,7 @@ def test_scale_array(count_matrix_format, zero_center):
     Test that running sc.pp.scale on an anndata object and an array returns the same results.
     """
     X = count_matrix_format(sp.random(100, 200, density=0.3).toarray())
-    adata = sc.AnnData(X=X.copy(), dtype=np.float64)
+    adata = sc.AnnData(X=X.copy())
 
     sc.pp.scale(adata, zero_center=zero_center)
     scaled_X = sc.pp.scale(X, zero_center=zero_center, copy=True)
@@ -257,7 +257,7 @@ def test_downsample_counts_per_cell(count_matrix_format, replace, dtype):
         * np.random.binomial(1, 0.3, (1000, 100))
     )
     X = X.astype(dtype)
-    adata = AnnData(X=count_matrix_format(X), dtype=dtype)
+    adata = AnnData(X=count_matrix_format(X).astype(dtype))
     with pytest.raises(ValueError):
         sc.pp.downsample_counts(
             adata, counts_per_cell=TARGET, total_counts=TARGET, replace=replace
@@ -293,7 +293,7 @@ def test_downsample_counts_per_cell_multiple_targets(
         * np.random.binomial(1, 0.3, (1000, 100))
     )
     X = X.astype(dtype)
-    adata = AnnData(X=count_matrix_format(X), dtype=dtype)
+    adata = AnnData(X=count_matrix_format(X).astype(dtype))
     initial_totals = np.ravel(adata.X.sum(axis=1))
     with pytest.raises(ValueError):
         sc.pp.downsample_counts(
@@ -324,7 +324,7 @@ def test_downsample_total_counts(count_matrix_format, replace, dtype):
         * np.random.binomial(1, 0.3, (1000, 100))
     )
     X = X.astype(dtype)
-    adata_orig = AnnData(X=count_matrix_format(X), dtype=dtype)
+    adata_orig = AnnData(X=count_matrix_format(X).astype(dtype))
     total = X.sum()
     target = np.floor_divide(total, 10)
     initial_totals = np.ravel(adata_orig.X.sum(axis=1))
